Bug: Extension now properly updates on site URL changes.
Before, the content script used location.href as a useEffect dependency
to handle Twitch's SPA behavior. This was insufficient, as going to a new
page did not re-trigger the content script. So, decided to use a
mutationObserver on the URL instead, as well as changing host permissions
and the 'matches' section of the manifest.json so that the content script
can run no matter where it is on the site and check on / listen to the URL.

Chore: Removed unused clsx package.